% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_dml.R
\name{att_dml}
\alias{att_dml}
\title{DML DDD estimator for ATT, with panel data and 2 periods}
\usage{
att_dml(did_preprocessed)
}
\arguments{
\item{did_preprocessed}{A list containing preprocessed data and specifications for the DDD estimation.
Expected elements include:
- \code{preprocessed_data}: A data table containing the data with variables needed for the analysis.
- \code{xformula}: The formula for the covariates to be included in the model. It should be of the form \code{~ x1 + x2}.
Default is \code{xformla = ~1} (no covariates).
- \code{learners}: A list of learners to be used in the estimation. It should be a list of two elements,
the first element being the learner for the propensity score and the second element being the learner for the outcome regression.
Default is \code{NULL}, then OLS and MLE Logit is used to estimate nuisances parameters. If \code{est_method = "dml"}, user have to specify \code{learners}.
- \code{n_folds}: The number of folds to be used in the cross-fitting. Default is \code{NULL}. If \code{est_method = "dml"}, user have to specify \code{n_folds}.
- \code{alpha}: The level of significance for the confidence intervals. Default is \code{0.05}.
- \code{subgroup_counts}: A matrix containing the number of observations in each subgroup.}
}
\value{
A list with the estimated ATT, standard error, upper and lower confidence intervals, and n_folds.
}
\description{
This function implements a double machine learning estimator for assessing the average
treatment effect on the treated (ATT) using a triple differences (DDD) approach
in panel data settings across two time periods. The function takes preprocessed
data structured specifically for this analysis.
}
\keyword{internal}
