% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nuisances.R
\name{compute_aggregation}
\alias{compute_aggregation}
\title{Compute Aggregated Treatment Effect Parameters}
\usage{
compute_aggregation(
  ddd_obj,
  type = "simple",
  balance_e = NULL,
  min_e = -Inf,
  max_e = Inf,
  na.rm = FALSE,
  boot = FALSE,
  nboot = NULL,
  cband = NULL
)
}
\arguments{
\item{call}{The function call to aggte}
}
\value{
Aggregation object (list) of class \code{\link{aggddd}}
}
\description{
Does the heavy lifting on computing aggregated group-time
average treatment effects
}
\keyword{internal}
