% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dml.R
\name{compute_dml_nuisances}
\alias{compute_dml_nuisances}
\title{Function to compute the DML ATT + score function}
\usage{
compute_dml_nuisances(
  data,
  xformula,
  ml_pa,
  ml_md,
  fold_test,
  fold_train,
  n_folds
)
}
\arguments{
\item{data}{A data.table containing the data processed}

\item{xformula}{The formula for the propensity score model}

\item{ml_pa}{\code{mlr3} learner for propensity score estimation}

\item{ml_md}{\code{mlr3} learner for regression adjustment estimation}

\item{fold_test}{List of test indices for global folds}

\item{fold_train}{List of train indices for global folds}

\item{n_folds}{The number of folds for cross-fitting}
}
\value{
A list containing ids, estimator for each k fold, and scores
}
\description{
Function to compute the DML ATT + score function
}
\keyword{internal}
