% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nuisances.R
\name{compute_scores}
\alias{compute_scores}
\title{Function to compute scores based on first step nuisances predictions}
\usage{
compute_scores(ml_pa, ml_md)
}
\arguments{
\item{ml_pa:}{MLR3 object for propensity score estimation}

\item{ml_md:}{MLR3 object for regression adjustment estimation}
}
\value{
: A list of scores and id for each k fold
}
\description{
Function to compute scores based on first step nuisances predictions
}
