% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fake_panel.R
\name{generate_test_panel}
\alias{generate_test_panel}
\title{Function to generate a fake dataset for testing}
\usage{
generate_test_panel(
  seed = 123,
  num_ids = 100,
  time = 2,
  initial.year = 2019,
  treatment.year = 2020
)
}
\arguments{
\item{seed}{Seed for reproducibility}

\item{num_ids}{Number of IDs}

\item{time}{Number of time periods}

\item{initial.year}{Initial year}

\item{treatment.year}{Treatment year}
}
\value{
A data.table with the following columns:
\itemize{
\item id: ID
\item state: State variable
\item year: Time variable
\item partition: Partition variable
\item x1: Covariate 1
\item x2: Covariate 2
\item treat: Treatment variable
\item outcome: Outcome variable
}
}
\description{
Function to generate fake dataset to test procedure
}
