% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dml.R
\name{get_local_folds}
\alias{get_local_folds}
\title{Function to get local sample indices for "local" DiDs comparisons.}
\usage{
get_local_folds(units_selected, n_folds, global_fold_test, global_fold_train)
}
\arguments{
\item{units_selected}{Vector of units selected for the local sample}

\item{n_folds}{Number of folds}

\item{global_fold_test}{List of test indices for global folds}

\item{global_fold_train}{List of train indices for global folds}
}
\value{
A list containing train and test indices for each fold at the local level.
}
\description{
Function to get local sample indices for "local" DiDs comparisons.
}
\keyword{internal}
