% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nuisances.R
\name{get_score_elements}
\alias{get_score_elements}
\title{Compute the score elements for the ATT}
\usage{
get_score_elements(y, d, p_hat, m_hat)
}
\arguments{
\item{y}{Outcome variable}

\item{d}{Treatment variable}

\item{p_hat}{Propensity score}

\item{m_hat}{Outcome regression}
}
\value{
A list containing the score elements psi_a and psi_b
}
\description{
Compute the score elements for the ATT
}
\keyword{internal}
