% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboot.R
\name{mboot}
\alias{mboot}
\title{Multiplier Bootstrap}
\usage{
mboot(inf_func, did_preprocessed, use_parallel = FALSE, cores = 1)
}
\arguments{
\item{inf_func}{an influence function}

\item{did_preprocessed}{A \code{dp} object obtained after preprocess}

\item{use_parallel}{Boolean of whether or not to use parallel processing in the multiplier
bootstrap, default is \code{use_parallel=FALSE}}

\item{cores}{the number of cores to use with parallel processing, default is \code{cores=1}}
}
\value{
A list with following elements
\item{bres}{results from each bootstrap iteration.}
\item{V}{variance matrix.}
\item{se}{standard errors.}
\item{crit_val}{a critical value for computing uniform confidence bands.}
}
\description{
This function take an influence function and use the
multiplier bootstrap to compute standard errors and critical values for
uniform confidence bands.
}
\keyword{internal}
