% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_attgt}
\alias{process_attgt}
\title{Process results inside att_gt_dr and att_gt_dml function}
\usage{
process_attgt(attgt_list)
}
\arguments{
\item{attgt_list}{A list of results from the \code{att_gt_dr} or \code{att_gt_dml} function.}
}
\value{
A list with three vectors: \code{group}, \code{att}, and \code{periods} containing the group, average treatment effect, and time periods respectively.
}
\description{
Process results inside att_gt_dr and att_gt_dml function
}
\keyword{internal}
