% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfe_ddd.R
\name{twfe_ddd}
\alias{twfe_ddd}
\title{TWFE DDD estimator, with panel data and 2 periods}
\usage{
twfe_ddd(yname, tname, dname, partition_name, xformla = ~1, data)
}
\arguments{
\item{yname}{A character string specifying the name of the outcome variable in the data.}

\item{tname}{A character string specifying the name of the time variable in the data.}

\item{dname}{A character string specifying the name of the state variable in the data.}

\item{partition_name}{A character string specifying the name of the partition variable in the data.}

\item{xformla}{A formula specifying the covariates to be included in the regression.}

\item{data}{A data frame containing the preprocessed data.}
}
\value{
A list with the estimated ATT, standard error, upper and lower confidence intervals, and influence function.
}
\description{
\code{twfe_ddd} function implements a two-way fixed effect regression for assessing the average
treatment effect on the treated (ATT) using a triple differences (DDD) specification
in panel data settings across two time periods.
}
