% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dml.R
\name{compute_dml_nuisances}
\alias{compute_dml_nuisances}
\title{Compute the propensity score for the given condition subgroup}
\usage{
compute_dml_nuisances(
  data,
  condition_subgroup,
  xformula,
  ml_pa,
  ml_md,
  n_folds
)
}
\arguments{
\item{data}{A data.table containing the data processed}

\item{condition_subgroup}{The condition subgroup for which to compute the propensity score}

\item{xformula}{The formula for the propensity score model}

\item{ml_pa}{The machine learning algorithm to use for the propensity score model}

\item{ml_md}{The machine learning algorithm to use for the regression adjustment model}

\item{n_folds}{The number of folds for cross-fitting}
}
\value{
A list containing ids, estimator for each k fold, and scores
}
\description{
Compute the propensity score for the given condition subgroup
}
\keyword{internal}
