% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nuisances.R
\name{compute_se_agg}
\alias{compute_se_agg}
\title{Take influence function and compute standard errors}
\usage{
compute_se_agg(influence_function, DIDparams = NULL)
}
\arguments{
\item{influence_function}{An influence function}
}
\value{
scalar standard error
}
\description{
Function to take an nx1 influence function and return
a standard error
}
\keyword{internal}
