% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_dgp_2periods}
\alias{gen_dgp_2periods}
\title{Function that generates panel data with staggered treatment assignment for 2 time periods}
\usage{
gen_dgp_2periods(size, dgp_type)
}
\arguments{
\item{size}{number of units}

\item{dgp_type}{type of DGP to generate.
1 if both nuisance functions are correct,
2 if only the outcome model is correct,
3 if only the pscore is correct,
4 if both nuisance functions are incorrect}
}
\value{
A list with the following elements:
\itemize{
\item data: data.table with the generated data with columns
- id: ID
- state: State variable
- time: Time variable
- partition: Partition variable
- x1: Covariate 1
- x2: Covariate 2
- x3: Covariate 3
- x4: Covariate 4
- y: Outcome variable
- cluster: Cluster variables (there's no actual within-cluster correlation)
\item att: True ATT. Set to be equal to 0.
\item att.unf: Computation of unfeasible ATT
\item eff: value of the theoretical efficiency bound
}
}
\description{
Function to generate data with staggered treatment adoption.
}
