% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_dgp_mult_periods}
\alias{gen_dgp_mult_periods}
\title{Function that generates panel data with staggered treatment assignment for multiple periods}
\usage{
gen_dgp_mult_periods(size, tperiods, dgp_type)
}
\arguments{
\item{size}{number of units}

\item{tperiods}{number of periods}

\item{dgp_type}{type of DGP to generate.
1 if both nuisance functions are correct,
2 if only the outcome model is correct,
3 if only the pscore is correct,
4 if both nuisance functions are incorrect}
}
\value{
A data.table with the following columns:
\itemize{
\item id: ID
\item state: State variable
\item year: Time variable
\item partition: Partition variable
\item x1: Covariate 1
\item x2: Covariate 2
\item treat: Treatment variable
\item outcome: Outcome variable
}
}
\description{
Function to generate data with staggered treatment adoption.
}
