% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_dgp_mult_periods}
\alias{gen_dgp_mult_periods}
\title{Function that generates panel data with staggered treatment assignment for multiple periods}
\usage{
gen_dgp_mult_periods(size, dgp_type = 1)
}
\arguments{
\item{size}{number of units}

\item{dgp_type}{type of DGP to generate.
1 if both nuisance functions are correct,
2 if only the outcome model is correct,
3 if only the pscore is correct,
4 if both nuisance functions are incorrect}
}
\value{
A list of 2 data.table with the following columns in long and wide format:
\itemize{
\item id: ID for panel data
\item cohort: Indicate the first period where the treatment is assigned
\item partition: Partition variable
\item x1: Covariate 1
\item x2: Covariate 2
\item x3: Covariate 3
\item x4: Covariate 4
\item cluster: Cluster variables (there's no actual within-cluster correlation)
\item time: Time periods
\item y: Outcome variable
}
}
\description{
Function to generate data with staggered treatment adoption.
Without loss of generality, the number of time periods is set to be 3.
}
